Move pages to their respective folders
