+++ dashboard/src/pages/dashboard/server/BattlesPage.tsx
