+++ dashboard/src/pages/dashboard/server/DeathsPage.tsx
