+++ dashboard/src/pages/dashboard/server/KillsPage.tsx
