+++ dashboard/src/routes.tsx
@@ -4,20 +4,20 @@ import AdminPage from "pages/admin/AdminPage";
 import AdminServersPage from "pages/admin/servers/AdminServersPage";
 import AdminSubscriptionPage from "pages/admin/subscriptions/AdminSubscriptionPage";
 import AdminSubscriptionsPage from "pages/admin/subscriptions/AdminSubscriptionsPage";
-import Auth from "pages/Auth";
-import AuthGuard from "pages/AuthGuard";
-import BattlesPage from "pages/BattlesPage";
-import DashboardPage from "pages/DashboardPage";
-import DeathsPage from "pages/DeathsPage";
-import HomePage from "pages/HomePage";
-import KillsPage from "pages/KillsPage";
-import PremiumPage from "pages/PremiumPage";
-import RankingsPage from "pages/RankingsPage";
-import ServerPage from "pages/ServerPage";
-import SettingsPage from "pages/SettingsPage";
-import SubscriptionPage from "pages/SubscriptionPage";
+import Auth from "pages/auth/Auth";
+import AuthGuard from "pages/auth/AuthGuard";
+import DashboardPage from "pages/dashboard/DashboardPage";
+import BattlesPage from "pages/dashboard/server/BattlesPage";
+import DeathsPage from "pages/dashboard/server/DeathsPage";
+import KillsPage from "pages/dashboard/server/KillsPage";
+import RankingsPage from "pages/dashboard/server/RankingsPage";
+import ServerPage from "pages/dashboard/server/ServerPage";
+import SettingsPage from "pages/dashboard/server/SettingsPage";
+import SubscriptionPage from "pages/dashboard/server/SubscriptionPage";
+import TrackPage from "pages/dashboard/server/TrackPage";
+import HomePage from "pages/home/HomePage";
+import PremiumPage from "pages/premium/PremiumPage";
 import SubscriptionsPage from "pages/subscriptions/SubscriptionsPage";
-import TrackPage from "pages/TrackPage";
 import { createRoutesFromElements, Navigate, Route } from "react-router-dom";
 
 const routes = createRoutesFromElements(
